pyproject.toml
[tool.poetry]
name = "pocdjangodeploy"
version = "0.1.0"
description = "POC Django Deploy on AWS."
authors = ["alfmorais <alfredneto@hotmail.com>"]
license = "MIT LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
Django = "^4.1.2"
django-jazzmin = "*"
djangorestframework = "^3.14.0"
pre-commit = "^2.20.0"
pytest = "7.2.0"
pytest-cov = "4.0.0"
pytest-django = "4.5.2"
python = "^3.10"
virtualenv = "20.0.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
